metrics: rename MetricsTemplate to DynMetricsTemplate
To denote that this is about templating dynamic counters.
