+++ lib/metrics/CMakeLists.txt
@@ -2,7 +2,7 @@ set(METRICS_HEADERS
     metrics/metrics.h
     metrics/dyn-metrics-store.h
     metrics/dyn-metrics-cache.h
-    metrics/metrics-template.h
+    metrics/dyn-metrics-template.h
     metrics/label-template.h
     PARENT_SCOPE)
 
@@ -10,7 +10,7 @@ set(METRICS_SOURCES
     metrics/metrics.c
     metrics/dyn-metrics-store.c
     metrics/dyn-metrics-cache.c
-    metrics/metrics-template.c
+    metrics/dyn-metrics-template.c
     metrics/label-template.c
     PARENT_SCOPE)
