+++ lib/metrics/Makefile.am
@@ -6,14 +6,14 @@ metricsinclude_HEADERS = \
 	lib/metrics/metrics.h	\
 	lib/metrics/dyn-metrics-store.h	\
 	lib/metrics/dyn-metrics-cache.h	\
-	lib/metrics/metrics-template.h	\
+	lib/metrics/dyn-metrics-template.h	\
 	lib/metrics/label-template.h
 
 metrics_sources = \
 	lib/metrics/metrics.c	\
 	lib/metrics/dyn-metrics-store.c	\
 	lib/metrics/dyn-metrics-cache.c	\
-	lib/metrics/metrics-template.c	\
+	lib/metrics/dyn-metrics-template.c	\
 	lib/metrics/label-template.c
 
 #include lib/metrics/tests/Makefile.am
