+++ lib/metrics/dyn-metrics-template.c
