+++ tests/copyright/policy
@@ -88,7 +88,7 @@ lib/cfg-monitor\.[ch]
 lib/cfg-source\.[ch]
 lib/metrics-pipe\.[ch]
 lib/metrics/label-template\.(c|h)$
-lib/metrics/metrics-template\.(c|h)$
+lib/metrics/dyn-metrics-template\.(c|h)$
 lib/metrics/dyn-metrics-store\.(c|h)$
 lib/metrics/dyn-metrics-cache\.(c|h)$
 lib/metrics/metrics\.(c|h)$
